requirements.txt
botocore==1.16.24
cftime==1.1.3
docutils==0.15.2
jmespath==0.10.0
netCDF4==1.5.3
numpy==1.18.5
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.9
